% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{bhattacharyya_metric}
\alias{bhattacharyya_metric}
\title{Bhattacharyya Distance Metric}
\usage{
bhattacharyya_metric()
}
\value{
A \code{distance_metric} object.
}
\description{
The Bhattacharyya distance between two groups' Gaussian posteriors.
}
