% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{initialize,PeriodicKernel-method}
\alias{initialize,PeriodicKernel-method}
\title{Initialize Method for PeriodicKernel}
\usage{
\S4method{initialize}{PeriodicKernel}(.Object, variance_per = 1, length_scale_per = 1, period = 1)
}
\arguments{
\item{.Object}{An object of class \code{PeriodicKernel}.}

\item{variance_per}{A numeric value for the signal variance. Defaults to a random uniform value between 0 and 3.}

\item{length_scale_per}{A numeric value for the length scale parameter. Defaults to a random uniform value between 0 and 3.}

\item{period}{A numeric value for the period. Defaults to a random uniform value between 0 and \(2\pi\).}
}
\value{
An initialized object of class \code{PeriodicKernel}.
}
\description{
Initializes an instance of the \code{PeriodicKernel} class with specified variance, length scale, and period parameters.
}
