% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{set_hyperparameters}
\alias{set_hyperparameters}
\title{Replace Hps Kernels}
\usage{
set_hyperparameters(kernel, values)
}
\arguments{
\item{kernel}{Previous kernel}

\item{values}{New HPs}
}
\value{
A kernel with new HPs
}
\description{
Update HPs of complex kernels
}
