% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{*,ANY,AbstractKernel-method}
\alias{*,ANY,AbstractKernel-method}
\title{Multiplication Method for AbstractKernel}
\usage{
\S4method{*}{ANY,AbstractKernel}(e1, e2)
}
\arguments{
\item{e1}{ANY.}

\item{e2}{An object of class AbstractKernel.}
}
\value{
An object of class AbstractKernel.
}
\description{
This method defines the multiplication operation for AbstractKernel objects.
}
