% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_diff.R
\name{calculate_group_overlaps}
\alias{calculate_group_overlaps}
\title{Compute Overlapping Coefficient}
\usage{
calculate_group_overlaps(results, n = 1000)
}
\arguments{
\item{results}{A list containing parameters of the posterior distribution for each group.}

\item{n}{A number indicating the number of samples generated during the overlapping coefficient computation process.}
}
\value{
A matrix with the coefficient for each group.
}
\description{
Use the bayestestR package to compute the overlapping coefficient.
}
