% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior.R
\name{multi_posterior_mean}
\alias{multi_posterior_mean}
\title{Compute Posterior Means for Different Groups}
\usage{
multi_posterior_mean(data, kern, mu_0 = 1, lambda_0 = 1)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed. Must include columns 'Group', 'ID', 'Output', and 'Input'.}

\item{kern}{A kernel function or object used to compute pairwise kernels.}

\item{mu_0}{Prior mean parameter.}

\item{lambda_0}{Prior precision parameter.}
}
\value{
A list of results for each group, containing the posterior means and kernel matrices.
}
\description{
This function computes the posterior means for different groups within a dataset.
It expects the input data frame to contain a column named 'Group' that identifies
the groups, as well as columns 'ID', 'Output', and 'Input'.
}
\details{
The posterior distribution is given by:

\eqn{p(\mathbf{\mu} \mid y_1, \dots, y_N, \Sigma_{\hat{\theta}}) = \mathcal{N}\left(\mathbf{\mu}; \ \dfrac{\lambda_0 \mu_0 + \sum_{n=1}^{N} y_n}{N + \lambda_0}, \dfrac{1}{N + \lambda_0} \Sigma_{\hat{\theta}}\right)}
}
