% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample from a Normal multivariate distribution}
\usage{
sample_posterior(results, n)
}
\arguments{
\item{results}{A list containing parameters of the posterior distribution for each group.
Each group should have elements \code{muk} (mean) and \code{sigmak} (covariance matrix). (from multi_posterior_mean)}

\item{n}{A number indicating the number of samples}
}
\value{
A list of sampled values for each group.
}
\description{
Sample
}
